Moved MarkdownHtmlConverter service into main app.
Last of the Pivotal-style "unbuilt Rails dependency" pseudo-Gems now
gone; no longer any need for the run-tests-and-statistics-on-each hack
used for much too long.

RSpec: 864 examples, 0 failures, 1 pending
RuboCop: 274 files inspected, no offenses detected
