--- lib/newpoc-services-markdown_html_converter/.rspec
@@ -1,2 +1,0 @@
---format documentation
---color
