--- lib/newpoc-services-markdown_html_converter/Gemfile
@@ -1,4 +1,0 @@
-source 'https://rubygems.org'
-
-# Specify your gem's dependencies in newpoc-services-markdown_html_converter.gemspec
-gemspec
