--- lib/newpoc-services-markdown_html_converter/lib/newpoc/services/markdown_html_converter/version.rb
@@ -1,8 +1,0 @@
-module Newpoc
-  module Services
-    # Convert possibly mixed Markdown/HTML content to HTML.
-    class MarkdownHtmlConverter
-      VERSION = '0.0.2'
-    end
-  end
-end
